user_service/requirements.txt
fastapi>=0.95.0
uvicorn>=0.21.1
pymongo>=4.3.3
python-multipart>=0.0.6
passlib>=1.7.4
python-jose>=3.3.0
email-validator>=2.0.0
python-dotenv>=1.0.0
passlib[argon2]